requirements.txt
cdktf
cdktf-cdktf-provider-azurerm
pydantic
pydantic-settings
python-dotenv
pyyaml